+++ spforge/cross_validator/cross_validator.py
@@ -19,13 +19,13 @@ class MatchKFoldCrossValidator(CrossValidator):
     """
 
     def __init__(
-            self,
-            match_id_column_name: str,
-            date_column_name: str,
-            predictor: BasePredictor,
-            scorer: Optional[BaseScorer] = None,
-            min_validation_date: Optional[str] = None,
-            n_splits: int = 3,
+        self,
+        match_id_column_name: str,
+        date_column_name: str,
+        predictor: BasePredictor,
+        scorer: Optional[BaseScorer] = None,
+        min_validation_date: Optional[str] = None,
+        n_splits: int = 3,
     ):
         """
         :param match_id_column_name: The column name of the match_id
@@ -44,10 +44,10 @@ class MatchKFoldCrossValidator(CrossValidator):
 
     @nw.narwhalify
     def generate_validation_df(
-            self,
-            df: FrameT,
-            return_features: bool = False,
-            add_train_prediction: bool = True,
+        self,
+        df: FrameT,
+        return_features: bool = False,
+        add_train_prediction: bool = True,
     ) -> IntoFrameT:
         """
         Generate predictions on validation dataset.
@@ -67,7 +67,8 @@ class MatchKFoldCrossValidator(CrossValidator):
 
             If set to false it will only return the predictions for the validation dataset
         """
-        if '__cv_row_index' not in df.columns:
+        ori_cols = df.columns
+        if "__cv_row_index" not in df.columns:
             df = df.with_row_index("__cv_row_index")
         if self.predictor.pred_column in df.columns:
             df = df.drop(self.predictor.pred_column)
@@ -79,7 +80,6 @@ class MatchKFoldCrossValidator(CrossValidator):
 
         predictor = copy.deepcopy(self.predictor)
         validation_dfs = []
-        ori_cols = df.columns
 
         if not self.min_validation_date:
             unique_dates = df[self.date_column_name].unique(maintain_order=True)
@@ -88,8 +88,8 @@ class MatchKFoldCrossValidator(CrossValidator):
 
         df = df.with_columns(
             (
-                    nw.col(self.match_id_column_name)
-                    != nw.col(self.match_id_column_name).shift(1)
+                nw.col(self.match_id_column_name)
+                != nw.col(self.match_id_column_name).shift(1)
             )
             .cum_sum()
             .fill_null(0)
@@ -99,7 +99,7 @@ class MatchKFoldCrossValidator(CrossValidator):
             df = df.with_columns(nw.col("__cv_match_number") + 1)
 
         if isinstance(self.min_validation_date, str) and df.schema.get(
-                self.date_column_name
+            self.date_column_name
         ) in (nw.Date, nw.Datetime):
             min_validation_date = datetime.strptime(
                 self.min_validation_date, "%Y-%m-%d"
@@ -180,16 +180,20 @@ class MatchKFoldCrossValidator(CrossValidator):
                 validation_df = df.filter(
                     (nw.col("__cv_match_number") >= train_cut_off_match_number)
                     & (
-                            nw.col("__cv_match_number")
-                            < train_cut_off_match_number + step_matches
+                        nw.col("__cv_match_number")
+                        < train_cut_off_match_number + step_matches
                     )
                 )
 
         concat_validation_df = nw.concat(validation_dfs).drop("__cv_match_number")
-        return_feats = list(set([*ori_cols, *predictor.columns_added,
-                        self.validation_column_name] + predictor.features if return_features else [*ori_cols,
-                                                                                                   *predictor.columns_added]))
-
+        return_feats = list(
+            set(
+                [*ori_cols, *predictor.columns_added, self.validation_column_name]
+                + predictor.features
+                if return_features
+                else [*ori_cols, *predictor.columns_added]
+            )
+        )
 
         return (
             concat_validation_df.unique(
@@ -197,6 +201,5 @@ class MatchKFoldCrossValidator(CrossValidator):
                 keep="first",
             )
             .sort("__cv_row_index")
-            .select(return_feats
-            )
+            .select(return_feats)
         )
