+++ spforge/transformers/lag_transformers/_lag.py
@@ -3,22 +3,20 @@ from typing import Optional
 
 import narwhals as nw
 from narwhals.typing import FrameT, IntoFrameT
-import pandas as pd
-import polars as pl
 
 from spforge import ColumnNames
-from spforge.transformers.base_transformer import (
-    BaseLagGenerator,
+
+from spforge.transformers.lag_transformers._utils import (
     required_lag_column_names,
     transformation_validator,
     historical_lag_transformations_wrapper,
     future_lag_transformations_wrapper,
     future_validator,
 )
-from spforge.utils import validate_sorting
+from spforge.transformers.lag_transformers import BaseLagTransformer
 
 
-class LagTransformer(BaseLagGenerator):
+class LagTransformer(BaseLagTransformer):
 
     def __init__(
         self,
