+++ tests/transformation/lag_transformers/test_opponent_rolling_mean.py
