+++ tests/transformation/lag_transformers/test_rolling_mean_binary.py
